Revert "latest version."
This reverts commit b09f79b359e0bbf616096b6ddfe56c3f3a867377.

# Conflicts:
#	pom.xml
#	src/main/java/HAFPIS/domain/SrchTaskBean.java
#	src/main/java/HAFPIS/service/FaceRecog.java
#	src/main/java/HAFPIS/service/FpRecog.java
#	src/main/java/HAFPIS/service/IrisRecog.java
#	src/main/java/HAFPIS/service/LatFpRecog.java
#	src/main/java/HAFPIS/service/LatPalmRecog.java
#	src/main/java/HAFPIS/service/PalmRecog.java

+++ src/main/java/HAFPIS/domain/FPTTRec.java
@@ -8,13 +8,13 @@ import java.io.Serializable;
 public class FPTTRec extends Rec<FPTTRec> implements Serializable {
 
     private static final long serialVersionUID = 1611229844832755558L;
-    public int[] rpscores;
-    public int[] fpscores;
+    public float[] rpscores;
+    public float[] fpscores;
 
     public FPTTRec() {
         position = 0;//在指纹查重中无用，置为0
-        rpscores = new int[10];
-        fpscores = new int[10];
+        rpscores = new float[10];
+        fpscores = new float[10];
     }
 
 }
