+++ src/main/java/HAFPIS/service/OneToF_FPTT.java
@@ -156,9 +156,9 @@ public class OneToF_FPTT extends Recog implements Runnable {
                             tempScore = temp;
                         }
                         if (j >= 0 && j < 10) {
-                            fpttRec.rpscores[j] = (int) temp;
+                            fpttRec.rpscores[j] = temp;
                         } else {
-                            fpttRec.fpscores[j - 10] = (int) temp;
+                            fpttRec.fpscores[j - 10] = temp;
                         }
                     }
                     fpttRec.score = tempScore;
